Disable FCHOWN for now on snapd
We will need to vendorize this component. On snapd, services are running as
root, and sqlite3 binding wants to fchown to drop ownership to euid/geuid.
This isn't allowed by the security policy for now (no mediation on that
syscall). Fork it (will vendorize and open a bug report on snapd).